requirements.txt
fastapi 
uvicorn 
pillow
tesseract
pytesseract
